暂存

+++ provider/group.go
@@ -88,7 +88,7 @@ func (srv *Service) QueryGroups(ctx context.Context, req *standard.QueryGroupsRe
 		return resp, nil
 	}
 
-	totalPage, currentPage, groups, err := dao.QueryGroups(int64(req.Page), int64(req.Limit))
+	totalPage, currentPage, groups, err := dao.QueryGroups(req.Page, req.Limit)
 	if err != nil {
 		resp.State = standard.State_DB_OPERATION_FATLURE
 		resp.Message = err.Error()
@@ -101,8 +101,8 @@ func (srv *Service) QueryGroups(ctx context.Context, req *standard.QueryGroupsRe
 	}
 
 	resp.State = standard.State_SUCCESS
-	resp.CurrentPage = int64(currentPage)
-	resp.TotalPage = int64(totalPage)
+	resp.CurrentPage = currentPage
+	resp.TotalPage = totalPage
 	resp.Message = "查询成功"
 	resp.Data = data
 	return resp, nil
